backend/catalogue-service/requirements.txt
fastapi
uvicorn
prisma
pydantic
python-jose